requirements.txt
streamlit
torch
torchvision
torchaudio
PyMuPDF
langchain-text-splitters
sentence-transformers
chromadb
google-generativeai
huggingface_hub